# Improves performance of hammingdist