Hello Liam, I have prepared a set of patches which improves performance of hammingdist use in the context of the covtrec project. You can review all the patches on https://github.com/ho-ob/hammingdist . If you are interested to get this patches into hammingdist I would create a pull request from there.

commit c86d0ae2254104fc243a915e8fbff1091b147ee0
This patch is non performance related. GitHub has changed there security policy and only allows git submodule checkout through secured connections.

commit fbf787ada34103fb53de3a6725c69ce3701bdca9
This patch changes the order in which the hardware features are checked. The most capable hardware feature is checked last so that it overwrites previously found less capable hardware features

commit ef63334a2f8e73970f7936af93c9d36ba06ec773
This patch adds a schedule directive for the OpenMP loop parallelization. Previously the scheduling was depending on the OpenMP runtime implementation. Normally this means, that the loop is divided in junks which a statically scheduled to the threads. This means a big load inbalance in our case, as the first thread always gets fewer work than the other threads because of the lower triangular shape of the matrix. The patch generates the best possible load balance by still keeping the overhead.

commit 886883e8ba24843102e37d536954ad74d47df92b
This patch speeds up generic distance computation. In my test case time went down from 1m12,219s to 0m46,584s. For comparision AVX2 implementation took 0m30,987s and SSE Implementation took 0m35,335s.

commit 74e3c91aeda77337c0f693759adc8a90ce59c310
This patch implements an OpenMP parallelized version of dump_lower_triangular. In a bigger testcase this bring execution time down from 61m42,280s to 16m40,637s. The non OpenMP version is not limited by the IO-subsystem but by the type conversion int -> string, which can be parallelized without overhead. Which adds an additional difficulty is the need to write the data ordered. This is the reason, why I used a static scheduling this time. So there are threads which finish with there work earlier but in sum even 4 threads are strong enough to hit the IO-limit.